search/requirements/base.txt
Django==1.10.6
django-cors-headers==2.0.0
djangorestframework==3.6.2
django-extensions==1.7.5
django-filter==1.0.1
mysqlclient==1.3.9
Pillow==4.0.0
pysolr==3.6.0
pytz==2016.10
pydocs==0.2
celery==4.0.2
mongoengine==0.11.0
mongomock==3.7.0
